backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
redis